Removed extensive logging data
"Removed extensive logging data

Introduced a recalculation of stoploss and takeprofit based on current configuration values. This update also includes a condition to check if the market is in position and has an entry price greater than zero. If these conditions are met, the expected stoploss and takeprofit are recalculated and updated if they differ significantly from their previous values."